Move most forgot_password logic to it's own lib directory.
