+++ app/libs/forgot_password/__init__.py
@@ -1,6 +1,6 @@
 from uuid import uuid4
 from app.models.user import User
-from app.models.forgot_password_token import ForgotPasswordToken
+from .model import ForgotPasswordToken
 
 
 class InvalidTokenException(Exception):
