+++ config/common.py
@@ -26,7 +26,11 @@ MODELS = ('user', 'post')
 
 # View blueprints to be loaded, in the order specified
 # See `load_blueprints` in `app/__init__.py` for more details
-BLUEPRINTS = ('main', {'name': 'admin', 'url_prefix': '/admin'})
+BLUEPRINTS = (
+    'main',
+    {'name': 'forgot_password', 'path': 'app.libs.forgot_password.views'},
+    {'name': 'admin', 'url_prefix': '/admin'}
+)
 
 # forgot password
 FORGOT_PASSWORD_TIMER = 3600
